Renamed overwrite_sink -> overwrite_channel.
